+++ packages/medical-domain/src/index.ts
@@ -50,7 +50,9 @@ import {
   type DeviceConfig,
   type ParameterConfig,
   type PumpConfig,
-  type ParametersChange
+  type ParametersChange,
+  type PumpSettingsParameter,
+  ChangeType
 } from './domains/models/medical/datum/pump-settings.model'
 import type PumpSettings from './domains/models/medical/datum/pump-settings.model'
 import type ReservoirChange from './domains/models/medical/datum/reservoir-change.model'
@@ -133,8 +135,10 @@ export {
   convertBG,
   GlycemiaStatisticsService,
   CarbsStatisticsService,
+  type ParametersChange,
+  type PumpSettingsParameter,
+  ChangeType,
   BasalBolusStatisticsService,
-  type BasicData,
-  type ParametersChange
+  type BasicData
 }
 export default MedicalDataService