+++ packages/medical-domain/test/mock/data.statistics.mock.ts
@@ -26,13 +26,14 @@
  */
 
 import type DateFilter from '../../src/domains/models/time/date-filter.model'
-import { createMealData, createWizardData, createRandomBasal, createRandomBolus } from '../data-generator'
-import { type Meal, type Wizard, type Basal, type Bolus } from '../../src'
+import { createMealData, createRandomBasal, createRandomBolus, createWizardData } from '../data-generator'
+import { type Basal, type Bolus, type Meal, type Wizard } from '../../src'
 
 const abbottDevice = 'AbbottFreeStyleLibre-XXX-XXXX'
 const dexcomDevice = 'Dexcom-XXX-XXXX'
 export const MS_IN_HOUR = 864e5 / 24
 
+export const MS_IN_HOUR = 864e5 / 24
 export const bgDataSourceOneDay: Array<[Date, string]> = [
   // data for one day and two days tests
   [new Date('2018-02-01T00:00:00.000Z'), abbottDevice],
@@ -94,6 +95,24 @@ export const buildWizardData = (data: Array<[Date, string]>): Wizard[] => (
     {
       ...createWizardData(wizardData[0]),
       deviceName: wizardData[1]
+    }
+  ))
+)
+export const buildBasalsData = (basalsData: Array<[Date, number, number, string]>): Basal[] => (
+  basalsData.map((basals) => (
+    {
+      ...createRandomBasal(basals[0], basals[2]),
+      rate: basals[1],
+      duration: basals[2],
+      deliveryType: basals[3]
+    }
+  ))
+)
+export const buildBolusData = (bolusData: Array<[Date, number]>): Bolus[] => (
+  bolusData.map((bolus) => (
+    {
+      ...createRandomBolus(bolus[0]),
+      normal: bolus[1]
     }
   ))
 )