+++ packages/yourloops/types/modules.d.ts
@@ -25,41 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import styles from '../diabeloop.css'
-import React, { type FunctionComponent } from 'react'
-import { ChangeType, type HistorizedParameter } from '../../../models/historized-parameter.model'
-import { useTranslation } from 'react-i18next'
-
-interface HistoryTableParameterChangeProps {
-  parameter: HistorizedParameter
-}
-
-const buildIconCssClass = (change: ChangeType): string => {
-  switch (change) {
-    case ChangeType.Added:
-      return 'icon-add'
-    case ChangeType.Deleted:
-      return 'icon-remove'
-    case ChangeType.Updated:
-      return 'icon-refresh'
-    default:
-      break
-  }
-  return 'icon-unsure-data'
-}
-
-export const HistoryTableParameterChange: FunctionComponent<HistoryTableParameterChangeProps> = (props): JSX.Element => {
-  const { t } = useTranslation('main')
-  const { parameter } = props
-
-  const iconClass = buildIconCssClass(parameter.changeType)
-  return (
-    <span>
-      <i className={iconClass} />
-      <span
-        className={styles.parameterHistory}>
-          {t(`params|${parameter.name}`)}
-      </span>
-    </span>
-  )
-}
+declare module '*.css'
+declare module '*.svg'
+declare module '*.png'
+declare module '*.otf'