+++ test/locales/translations.test.js
@@ -65,7 +65,11 @@ const ignoredTransKeysForBlip = [
   'delivered',
   'level',
   'wheel-label-off',
-  'wheel-label-on'
+  'wheel-label-on',
+  'Date',
+  'Level',
+  'Parameter',
+  'Parameters History'
 ]
 const ignoredTransKeyInBlipFiles = [
   '${physicalActivity.reportedIntensity}-pa',
@@ -196,6 +200,9 @@ const ignoredTransKeyForYourLoops = [
   'total-carbs',
   'hba1c',
   'no-medical-files-patient',
+  'added',
+  'deleted',
+  'updated',
   'avg-time-loop',
   'time-loop',
   'avg-time-loop-tooltip',
@@ -227,7 +234,19 @@ const ignoredTransKeyInYourLoopsFiles = [
   'params|${parameter.name}',
   'Device',
   'Unit',
-  'Value'
+  'Value',
+  'Cgm sensor expiration date',
+  'Cgm transmitter id',
+  'Cgm transmitter end of life',
+  'Cgm transmitter software version',
+  'Level',
+  'Parameter',
+  'Product',
+  'Pump',
+  'Pump cartridge expiration date',
+  'Pump version',
+  'Serial Number',
+  'Setting'
 ]
 
 /**