+++ CHANGELOG.md
@@ -49,6 +49,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2399 Cancel invite dialog new design
 - YLP-2430 Add checkbox to confirm that user is a hcp
 - YLP-2435 Remove "Telemonitored" filter on patient list
+- YLP-2447 Device page new design
 
 ### Engineering use
 - YLP-2044 Remove redux settings store